fix shitloads of stuff having health

+++ code/game/atom_health.dm
@@ -1,3 +1,5 @@
+/// If this atom uses the health system at all.
+/atom/var/should_use_health = FALSE
 /// The health of this atom. If this is null, it will set health to maxhealth on Initialize. Otherwise, you can set a custom health value to use at initialize.
 /atom/var/health
 /// The maximum health of this atom. If null, health is not used.
@@ -9,7 +11,7 @@
  * This proc is called to add damage to an atom. If there is no health left, it calls on_death().
  */
 /atom/proc/add_damage(damage, damage_flags, damage_type, armor_penetration, obj/weapon)
-	if(!damage || !maxhealth || (damage < 1))
+	if(!damage || !maxhealth || (damage < 1) || !should_use_health)
 		return FALSE
 
 	var/datum/component/armor/armor = GetComponent(/datum/component/armor)
@@ -34,6 +36,8 @@
  * For custom damage condition hints. Some structures may want different ones than the default, like the cult crystal.
  */
 /atom/proc/get_damage_condition_hints(mob/user, distance, is_adjacent)
+	if(!should_use_health)
+		return
 	if(health < maxhealth * 0.25)
 		. = SPAN_DANGER("\The [src] looks like it's about to break!")
 	else if(health < maxhealth * 0.5)
@@ -51,13 +55,15 @@
  * This proc is called by update_health() when the health of the atom hits zero. Handles the destruction of the atom, or you can override it to do different effects.
  */
 /atom/proc/on_death(damage, damage_flags, damage_type, armor_penetration, obj/weapon)
+	if(!should_use_health)
+		return
 	qdel(src)
 
 /**
  * This proc is called to set the atom's health directly.
  */
 /atom/proc/change_health(new_health)
-	if(!maxhealth)
+	if(!maxhealth || !should_use_health)
 		return
 
 	if(health >= maxhealth)
@@ -70,7 +76,7 @@
  * This proc is called to directly add to an atom's health (basically, to add it).
  */
 /atom/proc/add_health(repair_amount)
-	if(!maxhealth)
+	if(!maxhealth || !should_use_health)
 		return
 
 	if(health >= maxhealth)
